post: created database realization
Currently, we can create post and get last 3 posts from database
