+++ api/storage/models/post.go
@@ -1,17 +1,44 @@
 package models
 
-import "time"
+import (
+	"fmt"
+	"time"
+)
 
 type Post struct {
-	ID          int    `json:"ID"`
+	ID          int    `json:"omitempty"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
 	*Timestamp
 	*SoftDelete
 }
 
+func CreatePostWithData(title, desc string) *Post {
+	post := &Post{
+		Title:       title,
+		Description: desc,
+		Timestamp:   &Timestamp{},
+		SoftDelete:  &SoftDelete{},
+	}
+
+	post.SetTimestamp()
+	return post
+}
+
+func CreatePost() *Post {
+	post := &Post{
+		Title:       "",
+		Description: "",
+		Timestamp:   &Timestamp{},
+		SoftDelete:  &SoftDelete{},
+	}
+
+	post.SetTimestamp()
+	return post
+}
+
 func (p *Post) Validate() bool {
-	if len(p.Title) != 0 && len(p.Description) != 0 && p.ID != 0 {
+	if len(p.Title) != 0 && len(p.Description) != 0 {
 		return true
 	}
 
@@ -20,7 +47,7 @@ func (p *Post) Validate() bool {
 
 func (p *Post) SetTimestamp() {
 	p.UpdatedAt = time.Now().Format(time.RFC3339)
-	if len(p.CreatedAt) != 0 {
+	if len(p.CreatedAt) == 0 {
 		p.CreatedAt = time.Now().Format(time.RFC3339)
 	}
 }
@@ -37,7 +64,7 @@ func (m *Post) TableCreate() string {
 	return `
 		create table posts
 		(
-			id          serial
+			ID          serial
 				constraint posts_pk
 					primary key,
 			"title"  varchar(60)      not null,
@@ -47,4 +74,12 @@ func (m *Post) TableCreate() string {
 			"DeletedAt" timestamp
 		);
 	`
+}
+
+func (m *Post) CreateRecord() string {
+	return fmt.Sprintf(`INSERT INTO posts ("title", "description", "CreatedAt", "UpdatedAt", "DeletedAt") VALUES ('%s','%s','%s','%s', NULL)`, m.Title, m.Description, m.CreatedAt, m.UpdatedAt)
+}
+
+func GetLastPosts(limit int) string {
+	return fmt.Sprintf(`SELECT * FROM posts ORDER BY "CreatedAt" DESC LIMIT %d`, limit)
 }
