+++ api/v1/post.go
@@ -33,7 +33,9 @@ func (v *v1) GetLastPosts(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
+	w.Header().Set("content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
+
 	w.Write(response)
 }
